renamed src/api/apiConstants.js to src/api/config.js since it really is just pulling a configuration
