+++ src/components/NDExSave/assets/config.js
@@ -1,4 +1,4 @@
-import { SERVICE_SERVER_URL } from '../../../api/apiConstants'
+import { SERVICE_SERVER_URL } from '../../../api/config'
 
 const config = {
   save_to_ndex: SERVICE_SERVER_URL + '/network'
